ResumeIntelligenceSystem/requirements.txt
# Core dependencies
pymupdf==1.22.1
spacy==3.6.1
transformers==4.30.2
torch==2.0.1
python-docx==0.8.11
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
sentence-transformers==2.2.2
huggingface-hub==0.16.4

# Download spaCy model
# python -m spacy download en_core_web_lg